Onion
There are many other types of onions, including onions (Allium fistulosum). It is cultivated for foods such as onions (A. × proliferum) and Canadian onions (Allium canadense). The name "Wild Onion" is used for many types of green onions, but A. Known only from the development of sepia. The origin of his parents is still unknown, but there are rewards. [6] Onions are usually tall or perennial, but are usually thought to be first harvested during the first planting season.The onion has a turquoise depression, and after one day the onion swells at the edge of the onion. The bulbs have short, narrow beads surrounded by modified scales (ter և) that surround the central mound at the end of the stem. In the fall (or spring to winter) the owner dies and the outer layer of the bulb becomes dry and brittle. Crops are harvested, dried, prepared for use, or stored as onions. Bagel cultivation is associated with many pests and diseases, especially onions and garlic. A variety of fresh squid mushrooms ring. Onions: Some types of A. onions, such as potatoes, produce large